requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
httpx
apscheduler
python-dotenv
